pyproject.toml
[project]
name = "ai-agents-beyond-jupyter-notebook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "openai-agents[sqlalchemy]>=0.7.0",
    "psycopg2-binary>=2.9.11",
    "redis>=5.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
